# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tech-Admin-Api",
        "contact": {
            "name": "Tech-Admin",
            "url": "  http://tech-admin.com",
            "email": "  @gmail.com"
        },
        "version": "0.1"
    },
    "paths": {
...
        "/api/acesso/{id}": {
            "get": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Mostra um acesso",
                "description": "Mostra um acesso",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do acesso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o acesso"
                    },
                    "default": {
                        "description": "Caso não encontre o acesso"
                    }
                }
            },
            "put": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Atualiza um acesso",
                "description": "Atualiza um acesso",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do acesso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "description": "Nome do acesso",
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "description": "Descrição do acesso",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status do acesso",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o acesso atualizado"
                    },
                    "default": {
                        "description": "Caso não encontre o acesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Deleta um acesso",
                "description": "Deleta um acesso",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do acesso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o acesso deletado"
                    },
                    "default": {
                        "description": "Caso não encontre o acesso"
                    }
                }
            }
        },
...
        "/api/venda-produto": {
            "get": {
                "tags": [
                    "Carrinho"
                ],
                "summary": "Lista todos os produtos na venda",
                "description": "Retorna todos os produtos na venda",
                "operationId": "indexVendaProduto",
                "responses": {
                    "200": {
                        "description": "Retorna todos os produtos na venda"
                    },
                    "default": {
                        "description": "Caso não encontre o produto na venda"
                    }
                }
            },
            "post": {
                "tags": [
                    "Carrinho"
                ],
                "summary": "Cria um novo produto na venda",
                "description": "Cria um novo produto na venda",
                "operationId": "storeVendaProduto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_venda": {
                                        "description": "ID da venda",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "id_produto": {
                                        "description": "ID do produto",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "quantidade": {
                                        "description": "Quantidade do produto",
                                        "type": "integer",
                                        "format": "int64"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o produto na venda"
                    },
                    "default": {
                        "description": "Caso não encontre o produto na venda"
                    }
                }
            }
        },
        "/api/venda-produto/{id}": {
            "get": {
                "tags": [
                    "Carrinho"
                ],
                "summary": "Mostra um produto na venda",
                "description": "Mostra um produto na venda",
                "operationId": "showVendaProduto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto na venda",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o produto na venda"
                    },
                    "default": {
                        "description": "Caso não encontre o produto na venda"
                    }
                }
            },
            "put": {
                "tags": [
                    "Carrinho"
                ],
                "summary": "Atualiza um produto na venda",
                "description": "Atualiza um produto na venda",
                "operationId": "updateVendaProduto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto na venda",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_venda": {
                                        "description": "ID da venda",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "id_produto": {
                                        "description": "ID do produto",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "quantidade": {
                                        "description": "Quantidade do produto",
                                        "type": "integer",
                                        "format": "int64"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o produto na venda"
                    },
                    "default": {
                        "description": "Caso não encontre o produto na venda"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Carrinho"
                ],
                "summary": "Remove um produto na venda",
                "description": "Remove um produto na venda",
                "operationId": "destroyVendaProduto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto na venda",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o produto na venda"
                    },
                    "default": {
                        "description": "Caso não encontre o produto na venda"
                    }
                }
            }
        }
    }
}